+++ server/api/accessmap/dynamic.go
@@ -1,132 +1,136 @@
 package accessmap
 
 import (
-	"encoding/json"
-	"fmt"
-	"time"
-
 	"github.com/pkg/errors"
-	"github.com/seknox/trasa/server/api/groups"
-	"github.com/seknox/trasa/server/api/notif"
-	"github.com/seknox/trasa/server/api/policies"
 	"github.com/seknox/trasa/server/api/services"
 	"github.com/seknox/trasa/server/api/system"
+	"github.com/seknox/trasa/server/utils"
+	"net"
+	"strings"
+	"time"
+
+	"github.com/seknox/trasa/server/api/notif"
 	"github.com/seknox/trasa/server/api/users"
 	"github.com/seknox/trasa/server/consts"
 	"github.com/seknox/trasa/server/models"
-	"github.com/seknox/trasa/server/utils"
-	logger "github.com/sirupsen/logrus"
+	"github.com/sirupsen/logrus"
 )
 
-func CreateDynamicService(hostname, serviceType, userID, userEmail, privilege, orgID string) (*models.Service, error) {
-	sett, err := checkDynamicSettings(userID, orgID)
+func GetDynamicPolicy(groups []string, userID, orgID string) (*models.Policy, error) {
+	err := checkDynamicAccessSetting(orgID)
 	if err != nil {
 		return nil, err
 	}
-	service, err := createDynamicServiceAndAccessMap(hostname, serviceType, userID, userEmail, privilege, orgID, sett)
+	dPolicy, err := Store.GetDynamicAccessPolicy(groups, userID, orgID)
 	if err != nil {
-		return nil, err
+		return nil, errors.Errorf(`failed to get dynamic access policy: %v`, err)
 	}
-	return service, err
+
+	return dPolicy, nil
 }
 
-func CreateDynamicAccessMap(serviceID, userID, userEmail, privilege, orgID string) (*models.Policy, error) {
-	sett, err := checkDynamicSettings(userID, orgID)
+func CreateDynamicService(hostname, serviceType, userEmail, orgID string) (*models.Service, error) {
+	err := checkDynamicAccessSetting(orgID)
 	if err != nil {
 		return nil, err
 	}
-	err = Store.CreateServiceUserMap(&models.ServiceUserMap{
-		MapID:     utils.GetUUID(),
-		ServiceID: serviceID,
-		OrgID:     orgID,
-		UserID:    userID,
-		PolicyID:  sett.PolicyID,
-		Privilege: privilege,
-		AddedAt:   time.Now().Unix(),
-	})
 
+	newService := NewService(hostname, serviceType, orgID)
+
+	err = checkService(newService)
 	if err != nil {
-		return nil, errors.WithMessage(err, "create new dynamic access map")
+		return nil, errors.WithMessage(err, "could not connect to service")
 	}
 
-	policy, err := policies.Store.GetPolicy(sett.PolicyID, orgID)
-	if err != nil {
-		logger.Errorf("get dynamic access policy: %v", err)
-		return nil, err
-	}
-
-	service, err := services.Store.GetFromID(serviceID)
-	if err != nil {
-		return &policy, err
-	}
-
-	notifyAdmins(orgID, fmt.Sprintf("%s (%s)", service.Name, service.Hostname), service.Type, userEmail)
-
-	return &policy, nil
-}
-
-func checkDynamicSettings(userID, orgID string) (*models.GlobalDynamicAccessSettings, error) {
-
-	//Check global settings
-	sett, err := system.Store.GetGlobalSetting(orgID, consts.GLOBAL_DYNAMIC_ACCESS)
-	if err != nil || !sett.Status {
-		logger.Error(err)
-		return nil, errors.Errorf("trasa: dynamic auth app not set: %v", err)
-	}
-
-	var settingVal models.GlobalDynamicAccessSettings
-	err = json.Unmarshal([]byte(sett.SettingValue), &settingVal)
-	if err != nil {
-		logger.Error(err)
-		return nil, errors.WithMessage(err, "trasa: invalid global settings")
-	}
-
-	check, err := groups.Store.CheckIfUserInGroup(userID, orgID, settingVal.UserGroups)
-	if err != nil || !check {
-		logger.Error(err)
-		return &settingVal, errors.Errorf("trasa: user cannot access dynamic auth app: %v", err)
-	}
-
-	return &settingVal, nil
-}
-
-func createDynamicServiceAndAccessMap(hostname, serviceType, userID, userEmail, privilege, orgID string, sett *models.GlobalDynamicAccessSettings) (*models.Service, error) {
-
-	newservice := NewService(hostname, serviceType, orgID)
-
-	err := services.Store.Create(&newservice)
+	err = services.Store.Create(&newService)
 	if err != nil {
 		return nil, errors.WithMessage(err, "create new dynamic service")
 	}
 
-	err = Store.CreateServiceUserMap(&models.ServiceUserMap{
-		MapID:     utils.GetUUID(),
-		ServiceID: newservice.ID,
-		OrgID:     orgID,
-		UserID:    userID,
-		PolicyID:  sett.PolicyID,
-		Privilege: privilege,
-		AddedAt:   time.Now().Unix(),
-	})
-
-	if err != nil {
-		return nil, errors.WithMessage(err, "create new dynamic access map")
-	}
-
 	notifyAdmins(orgID, hostname, serviceType, userEmail)
 
-	return &newservice, err
+	return &newService, err
+}
+
+func checkService(s models.Service) error {
+	host := s.Hostname
+	if !strings.Contains(s.Hostname, ":") {
+		switch s.Type {
+		case "ssh":
+			host = host + ":22"
+		case "rdp":
+			host = host + ":3389"
+		}
+	}
+
+	c, err := net.Dial("tcp", host)
+	if err != nil {
+		return err
+	}
+	return c.Close()
+}
+
+func checkDynamicAccessSetting(orgID string) error {
+	//Check global settings
+	sett, err := system.Store.GetGlobalSetting(orgID, consts.GLOBAL_DYNAMIC_ACCESS)
+	if err != nil {
+		logrus.Error(err)
+		return errors.Errorf("trasa: could not get dynamic access settings: %v", err)
+
+	}
+	if !sett.Status {
+		return errors.New("trasa: dynamic access disabled")
+	}
+	return nil
+}
+
+//NewService returns empty struct of service
+func NewService(hostname, stype, orgID string) models.Service {
+	return models.Service{
+		ID:              utils.GetUUID(),
+		OrgID:           orgID,
+		Name:            hostname,
+		SecretKey:       utils.GetRandomString(17),
+		Passthru:        false,
+		Hostname:        hostname,
+		Type:            stype,
+		ManagedAccounts: "",
+		RemoteAppName:   "",
+		Adhoc:           false,
+		NativeLog:       false,
+		RdpProtocol:     "nla",
+		ProxyConfig: models.ReverseProxy{
+			RouteRule:           "",
+			PassHostheader:      false,
+			UpstreamServer:      "",
+			StrictTLSValidation: true,
+		},
+		PublicKey:             "",
+		ExternalProviderName:  "",
+		ExternalID:            "",
+		ExternalSecurityGroup: "{}",
+		DistroName:            "",
+		DistroVersion:         "",
+		IPDetails: models.IPDetails{
+			IpAddress:      "0.0.0.0",
+			NetMask:        "",
+			DefaultGateway: "0.0.0.0",
+		},
+		CreatedAt: time.Now().Unix(),
+		UpdatedAt: time.Now().Unix(),
+		DeletedAt: 0,
+	}
 }
 
 func notifyAdmins(orgID, hostname, appType, userEmail string) {
 
 	admins, err := users.Store.GetAdminEmails(orgID)
 	if err != nil {
-		logger.Errorf("get admin email: %v", err)
+		logrus.Errorf("get admin email: %v", err)
 	}
 
 	if len(admins) < 1 {
-		logger.Error("no admins found")
+		logrus.Error("no admins found")
 		return
 	}
 
@@ -139,7 +143,7 @@ func notifyAdmins(orgID, hostname, appType, userEmail string) {
 	tmplt.TimeInt = time.Now().Unix()
 	err = notif.Store.SendEmail(orgID, consts.EMAIL_DYNAMIC_ACCESS, tmplt)
 	if err != nil {
-		logger.Errorf("send dynamic access email: %v", err)
+		logrus.Errorf("send dynamic access email: %v", err)
 	}
 
 }
